* 扫描收集所有模板，解析模板中标签
* 通过模板中的记号更新模板间关系

* 指定模板名称，根据关系树生成完整的HTML（不包括动态成分如变量等）
