+++ llt/ut/res/test_main.mod
@@ -1,11 +1,11 @@
-{name main_page_mod}
+{{name main_page_mod}}
 <html>
     <head>
         <title>This is a test page</title>
     </head>
     <body>
         <div>Hello</div>
-        {ac}
+        {{ac}}
     </body>
 </html>
