+++ llt/ut/res/test_ref_a.mod
@@ -1,4 +1,4 @@
-{name ref_a}
+{{name ref_a}}
 <div>
     A
 </div>
