+++ llt/ut/res/test_ref_b.mod
@@ -1,4 +1,4 @@
-{name ref_b}
+{{name ref_b}}
 <div>
     B
 </div>
